Split remaining LLVM op emission categories
