+++ crates/dialect-llvm/src/export/op_emission/mod.rs
@@ -8,10 +8,7 @@ use std::fmt::Write;
 
 use pliron::{
     basic_block::BasicBlock,
-    builtin::{
-        attributes::{FPDoubleAttr, FPSingleAttr, IntegerAttr},
-        op_interfaces::{CallOpCallable, CallOpInterface},
-    },
+    builtin::op_interfaces::{CallOpCallable, CallOpInterface},
     context::Ptr,
     op::Op,
     operation::Operation,
@@ -20,16 +17,19 @@ use pliron::{
 };
 
 use crate::{
-    attributes::{FPHalfAttr, ICmpPredicateAttr},
     ops,
     types::{FuncType, VoidType},
 };
 
-use super::{ModuleExportState, format_float_literal, format_half_literal, strip_device_prefix};
+use super::{ModuleExportState, strip_device_prefix};
 
+mod aggregate;
+mod arithmetic;
 mod atomic;
+mod casts;
 mod memory;
 mod terminator;
+mod virtual_ops;
 
 impl<'a> ModuleExportState<'a> {
     pub(super) fn export_op(
@@ -142,15 +142,7 @@ impl<'a> ModuleExportState<'a> {
                 self.export_binop("frem", op, value_names, output)?;
             }
             id if id == ops::FNegOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let arg = op_ref.get_operand(0);
-
-                write!(output, "  {res_name} = fneg ").unwrap();
-                self.export_type(arg.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(arg, value_names, output)?;
-                writeln!(output).unwrap();
+                self.export_fneg_op(&op_ref, value_names, output)?;
             }
             id if id == ops::SDivOp::get_opid_static() => {
                 self.export_binop("sdiv", op, value_names, output)?;
@@ -183,68 +175,12 @@ impl<'a> ModuleExportState<'a> {
                 self.export_binop("or", op, value_names, output)?;
             }
             id if id == ops::ICmpOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let lhs = op_ref.get_operand(0);
-                let rhs = op_ref.get_operand(1);
-
                 let icmp = op_obj.as_ref().downcast_ref::<ops::ICmpOp>().unwrap();
-                let pred_attr = icmp.predicate(self.ctx);
-                let pred_str = match pred_attr {
-                    ICmpPredicateAttr::EQ => "eq",
-                    ICmpPredicateAttr::NE => "ne",
-                    ICmpPredicateAttr::SLT => "slt",
-                    ICmpPredicateAttr::SLE => "sle",
-                    ICmpPredicateAttr::SGT => "sgt",
-                    ICmpPredicateAttr::SGE => "sge",
-                    ICmpPredicateAttr::ULT => "ult",
-                    ICmpPredicateAttr::ULE => "ule",
-                    ICmpPredicateAttr::UGT => "ugt",
-                    ICmpPredicateAttr::UGE => "uge",
-                };
-
-                write!(output, "  {res_name} = icmp {pred_str} ").unwrap();
-                self.export_type(lhs.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(lhs, value_names, output)?;
-                write!(output, ", ").unwrap();
-                self.export_value(rhs, value_names, output)?;
-                writeln!(output).unwrap();
+                self.export_icmp_op(&op_ref, icmp, value_names, output)?;
             }
             id if id == ops::FCmpOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let lhs = op_ref.get_operand(0);
-                let rhs = op_ref.get_operand(1);
-
                 let fcmp = op_obj.as_ref().downcast_ref::<ops::FCmpOp>().unwrap();
-                let pred_attr = fcmp.predicate(self.ctx);
-                let pred_str = match pred_attr {
-                    crate::attributes::FCmpPredicateAttr::False => "false",
-                    crate::attributes::FCmpPredicateAttr::OEQ => "oeq",
-                    crate::attributes::FCmpPredicateAttr::OGT => "ogt",
-                    crate::attributes::FCmpPredicateAttr::OGE => "oge",
-                    crate::attributes::FCmpPredicateAttr::OLT => "olt",
-                    crate::attributes::FCmpPredicateAttr::OLE => "ole",
-                    crate::attributes::FCmpPredicateAttr::ONE => "one",
-                    crate::attributes::FCmpPredicateAttr::ORD => "ord",
-                    crate::attributes::FCmpPredicateAttr::UEQ => "ueq",
-                    crate::attributes::FCmpPredicateAttr::UGT => "ugt",
-                    crate::attributes::FCmpPredicateAttr::UGE => "uge",
-                    crate::attributes::FCmpPredicateAttr::ULT => "ult",
-                    crate::attributes::FCmpPredicateAttr::ULE => "ule",
-                    crate::attributes::FCmpPredicateAttr::UNE => "une",
-                    crate::attributes::FCmpPredicateAttr::UNO => "uno",
-                    crate::attributes::FCmpPredicateAttr::True => "true",
-                };
-
-                write!(output, "  {res_name} = fcmp {pred_str} ").unwrap();
-                self.export_type(lhs.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(lhs, value_names, output)?;
-                write!(output, ", ").unwrap();
-                self.export_value(rhs, value_names, output)?;
-                writeln!(output).unwrap();
+                self.export_fcmp_op(&op_ref, fcmp, value_names, output)?;
             }
 
             // --- Calls ---
@@ -470,35 +406,8 @@ impl<'a> ModuleExportState<'a> {
                 self.export_cast("addrspacecast", op, value_names, output)?;
             }
             id if id == ops::ZExtOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let val = op_ref.get_operand(0);
-
                 let zext = op_obj.as_ref().downcast_ref::<ops::ZExtOp>().unwrap();
-                // Manual attribute access since helper is missing
-                let nneg_key: pliron::identifier::Identifier = "llvm_nneg_flag".try_into().unwrap();
-                let nneg = zext
-                    .get_operation()
-                    .deref(self.ctx)
-                    .attributes
-                    .0
-                    .get(&nneg_key)
-                    .and_then(|attr| {
-                        attr.downcast_ref::<pliron::builtin::attributes::BoolAttr>()
-                            .map(|b| bool::from(b.clone()))
-                    })
-                    .unwrap_or(false);
-
-                write!(output, "  {res_name} = zext ").unwrap();
-                if nneg {
-                    write!(output, "nneg ").unwrap();
-                }
-                self.export_type(val.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(val, value_names, output)?;
-                write!(output, " to ").unwrap();
-                self.export_type(res.get_type(self.ctx), output)?;
-                writeln!(output).unwrap();
+                self.export_zext_op(&op_ref, zext, value_names, output)?;
             }
             id if id == ops::SExtOp::get_opid_static() => {
                 self.export_cast("sext", op, value_names, output)?;
@@ -531,105 +440,32 @@ impl<'a> ModuleExportState<'a> {
                 self.export_cast("fptrunc", op, value_names, output)?;
             }
             id if id == ops::UndefOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                value_names.insert(res, "undef".to_string());
+                self.export_undef_op(&op_ref, value_names);
             }
 
             // --- Aggregate Ops ---
             id if id == ops::ExtractValueOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let agg = op_ref.get_operand(0);
-
                 let extract_op = op_obj
                     .as_ref()
                     .downcast_ref::<ops::ExtractValueOp>()
                     .unwrap();
-                let indices = extract_op.indices(self.ctx);
-
-                write!(output, "  {res_name} = extractvalue ").unwrap();
-                self.export_type(agg.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(agg, value_names, output)?;
-                for idx in indices {
-                    write!(output, ", {idx}").unwrap();
-                }
-                writeln!(output).unwrap();
+                self.export_extract_value_op(&op_ref, extract_op, value_names, output)?;
             }
             id if id == ops::InsertValueOp::get_opid_static() => {
-                let res = op_ref.get_result(0);
-                let res_name = value_names.get(&res).unwrap();
-                let agg = op_ref.get_operand(0);
-                let val = op_ref.get_operand(1);
-
                 let insert_op = op_obj
                     .as_ref()
                     .downcast_ref::<ops::InsertValueOp>()
                     .unwrap();
-                let indices = insert_op.indices(self.ctx);
-
-                write!(output, "  {res_name} = insertvalue ").unwrap();
-                self.export_type(agg.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(agg, value_names, output)?;
-                write!(output, ", ").unwrap();
-                self.export_type(val.get_type(self.ctx), output)?;
-                write!(output, " ").unwrap();
-                self.export_value(val, value_names, output)?;
-
-                for idx in indices {
-                    write!(output, ", {idx}").unwrap();
-                }
-                writeln!(output).unwrap();
+                self.export_insert_value_op(&op_ref, insert_op, value_names, output)?;
             }
 
             // --- Address Operations ---
             id if id == ops::AddressOfOp::get_opid_static() => {
-                // AddressOfOp is virtual in textual LLVM IR: every use site
-                // prints the global symbol directly. The naming pre-pass in
-                // export_func registers the result as `@<global_name>` before
-                // any block is emitted, so there is nothing to write here.
-                // The debug-only assertion keeps the contract honest if the
-                // pre-pass is ever refactored.
-                let res = op_ref.get_result(0);
-                debug_assert!(
-                    value_names
-                        .get(&res)
-                        .is_some_and(|name| name.starts_with('@')),
-                    "AddressOfOp result must be pre-registered as a global \
-                     symbol by the naming pre-pass; got {:?}",
-                    value_names.get(&res),
-                );
+                self.check_address_of_op(&op_ref, value_names);
             }
-
-            // --- Constants (Virtual) ---
             id if id == ops::ConstantOp::get_opid_static() => {
                 let const_op = op_obj.as_ref().downcast_ref::<ops::ConstantOp>().unwrap();
-                let val_attr = const_op.get_value(self.ctx);
-
-                let const_str = if let Some(int_attr) = val_attr.downcast_ref::<IntegerAttr>() {
-                    // Use APInt's proper decimal string conversion instead of parsing debug format.
-                    // The old code parsed debug strings like "APInt { value: 0x4000_0000_0000_u64 }"
-                    // by splitting on '_', which broke for values with underscore grouping
-                    // (e.g., 1u64 << 46 = 0x4000_0000_0000 would become 0x4000 = 16384).
-                    int_attr.value().to_string_unsigned_decimal()
-                } else if let Some(fp16_attr) = val_attr.downcast_ref::<FPHalfAttr>() {
-                    format_half_literal(fp16_attr.to_bits())
-                } else if let Some(fp32_attr) = val_attr.downcast_ref::<FPSingleAttr>() {
-                    // Float constant (f32) - format as LLVM float literal
-                    let float_val: f32 = fp32_attr.clone().into();
-                    format_float_literal(f64::from(float_val))
-                } else if let Some(fp64_attr) = val_attr.downcast_ref::<FPDoubleAttr>() {
-                    // Double constant (f64) - format as LLVM float literal
-                    let float_val: f64 = fp64_attr.clone().into();
-                    format_float_literal(float_val)
-                } else {
-                    "0".to_string() // Fallback
-                };
-
-                // Overwrite register name with constant literal
-                let res = op_ref.get_result(0);
-                value_names.insert(res, const_str);
+                self.export_constant_op(&op_ref, const_op, value_names);
             }
 
             // --- Unknown op fallback ---
